% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/falcons.R
\docType{data}
\name{falcons}
\alias{falcons}
\title{Falcons data}
\format{A data frame with 4 columns
\describe{
  \item{Year}{Year of count}
  \item{Pairs}{Number of adult pairs}
  \item{R.Pairs}{Number of reproductive pairs}
  \item{Eyasses}{Total number of fledged young from all reproductive nests}
}}
\source{
\url{https://www.vogelwarte.ch/de/projekte/publikationen/bpa/complete-code-and-data-files-of-the-book}
}
\usage{
falcons
}
\description{
Annual counts of the peregrine falcon population nesting in the French Jura from 1964 to 2003.
}
\keyword{datasets}
