% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WILD6900-package.R
\docType{package}
\name{WILD6900-package}
\alias{WILD6900}
\alias{WILD6900-package}
\title{WILD6900: Course materials for WILD6900: Applied Bayesian Analysis for Ecological Data}
\description{
Course materials for WILD6900: Applied Bayesian Analysis for Ecological Data at Utah State University. This package primarily exists to distribute data, code, and other materials in an efficient manner.
}
\author{
\strong{Maintainer}: Clark Rushing \email{clark.rushing@usu.edu}

}
\keyword{internal}
