% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/falcons.R
\docType{data}
\name{falcons}
\alias{falcons}
\title{Falcons data}
\format{A data frame with 4 columns
\describe{
  \item{Year}{Year of count}
  \item{Pairs}{YNumber of breeding pairs}
  \item{R.Pairs}{Year of count}
  \item{Eyasses}{Number of fledglings}
}}
\source{
\url{https://www.vogelwarte.ch/de/projekte/publikationen/bpa/complete-code-and-data-files-of-the-book}
}
\usage{
falcons
}
\description{
A dataset containing counts of Perigrine Falcons.
}
\keyword{datasets}
