% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WILD6900-package.R
\docType{package}
\name{WILD6900-package}
\alias{WILD6900}
\alias{WILD6900-package}
\title{WILD6900: Course materials for WILD6900: Applied Bayesian Analysis for Ecological Data}
\description{
More about what it does (maybe more than one line)
    Use four spaces when indenting paragraphs within the Description.
}
\author{
\strong{Maintainer}: Clark Rushing \email{clark.rushing@usu.edu}

}
\keyword{internal}
